+++ apps/chaingraph-frontend/src/store/trpc/store.ts
@@ -1,0 +1,25 @@
+/*
+ * Copyright (c) 2025 BadLabs
+ *
+ * Use of this software is governed by the Business Source License 1.1 included in the file LICENSE.txt.
+ *
+ * As of the Change Date specified in that file, in accordance with the Business Source License, use of this software will be governed by the Apache License, version 2.0.
+ */
+
+import type { TRPCClient } from '@badaitech/chaingraph-trpc/client'
+import { createTRPCClient } from '@badaitech/chaingraph-trpc/client'
+import { createEvent, createStore } from 'effector'
+import SuperJSON from 'superjson'
+
+export const createTRPCClientEvent = createEvent<{
+  trpcURL?: string
+  superjsonCustom?: typeof SuperJSON
+}>()
+
+export const $trpcClient = createStore<TRPCClient | null>(null)
+  .on(createTRPCClientEvent, (_, { trpcURL, superjsonCustom }) => {
+    return createTRPCClient({
+      url: trpcURL ?? `ws://localhost:3001`,
+      superjsonCustom: superjsonCustom ?? SuperJSON,
+    })
+  })