+++ apps/chaingraph-frontend/src/trpc/client.ts
@@ -6,30 +6,34 @@
  * As of the Change Date specified in that file, in accordance with the Business Source License, use of this software will be governed by the Apache License, version 2.0.
  */
 
-import { createTRPCClient } from '@badaitech/chaingraph-trpc/client'
-import SuperJSON from 'superjson'
-
 // export const _ = createTRPCClient({
 //   // TODO: from env
 //   url: `ws://localhost:3001`,
 //   superjsonCustom: SuperJSON,
 // })
 
-type TRPCClient = ReturnType<typeof createTRPCClient>
+// export type TRPCClient = ReturnType<typeof createTRPCClient>
 
-let _trpcClient: TRPCClient | null = null
-
-export function getStaticTRPCClient() {
-  if (_trpcClient === null) {
-    _trpcClient = createTRPCClient({
-      url: `ws://localhost:3001`,
-      superjsonCustom: SuperJSON,
-    })
-  }
-
-  return _trpcClient
-}
-
-export function setStaticTRPCClient(client: TRPCClient) {
-  _trpcClient = client
-}
+// let _trpcClient: TRPCClient | null = null
+//
+// export function getStaticTRPCClient(
+//   trpcURL?: string,
+//   superjsonCustom?: typeof SuperJSON,
+// ) {
+//   if (_trpcClient === null) {
+//     _trpcClient = createTRPCClient({
+//       url: trpcURL ?? `ws://localhost:3001`,
+//       superjsonCustom,
+//     })
+//   }
+//
+//   return _trpcClient
+// }
+//
+// export function isStaticTRPCClientExists() {
+//   return !!_trpcClient
+// }
+//
+// export function setStaticTRPCClient(client: TRPCClient) {
+//   _trpcClient = client
+// }